% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_concepts}
\alias{get_concepts}
\title{Get concepts from Alchemy API}
\usage{
get_concepts(url, api_key)
}
\arguments{
\item{url}{URL to extract the concepts from}

\item{api_key}{}
}
\value{
Dataframe with a concept for each word, the corrisponding score
and the origional url
}
\description{
Get concepts from Alchemy API
}
\examples{
\dontrun{
get_concepts('google.com', '1234')
}
}

