% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_emotions}
\alias{get_emotions}
\title{Get emotions from Alchemy API}
\usage{
get_emotions(url, api_key)
}
\arguments{
\item{url}{URL to extract the emotions from}

\item{api_key}{}
}
\value{
Dataframe with a row for each emotion, the score for that emotion
And the origional URL.
}
\description{
Get emotions from Alchemy API
}
\examples{
\dontrun{
get_emotions('google.com', '1234')
}
}

