% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_sentiment}
\alias{get_sentiment}
\title{Get sentiment from Alchemy API}
\usage{
get_sentiment(url, api_key)
}
\arguments{
\item{url}{URL to extract sentiment from}

\item{api_key}{}
}
\value{
Dataframe with a row for each emotion, the score for that emotion
And the origional URL.
}
\description{
Get sentiment from Alchemy API
}
\examples{
\dontrun{
get_sentiment('google.com', '1234')
}
}

