% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_text}
\alias{get_text}
\title{Get raw text of a URL}
\usage{
get_text(url, api_key)
}
\arguments{
\item{url}{URL to extract the text from}

\item{api_key}{}
}
\value{
One element vector with page's text
}
\description{
Get raw text of a URL
}
\examples{
\dontrun{
get_text('google.com', '1234')
}

}

